fix bugs after revert

+++ core/src/com/mikroysoft/game/ProgressBar.java
@@ -4,8 +4,8 @@ import com.badlogic.gdx.graphics.Texture;
 
 public class ProgressBar {
     public Texture texture;
-    private int height;
-    private int width;
+    private float height;
+    private float width;
     public Coordinate position;
     private int maximum;
     private int current;
@@ -27,23 +27,23 @@ public class ProgressBar {
         barFill = 0;
     }
 
-    public int getHeight () {
+    public float getHeight () {
         return this.height;
     }
 
     public int getFill() {
         return this.barFill;
     }
-    public int getWidth () {
+    public float getWidth () {
         return this.width;
     }
 
-    public void setDimensions(int w, int h) {
+    public void setDimensions(float w, float h) {
         this.height = h;
         this.width = w;
     }
 
-    public void setPosition(int x, int y) {
+    public void setPosition(float x, float y) {
         this.position = new Coordinate(x, y);
     }
